requirements.txt
Django==5.0.4
djangorestframework==3.14.0
psycopg2-binary==2.9.9
Pillow==10.3.0
python-decouple==3.8
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
django-extensions==3.2.3
openpyxl==3.1.2
reportlab==4.0.9

